Fix lots of test failures due to Go 1.7 upgrade, mostly minor type issues

+++ inspeqtor_test.go
@@ -20,7 +20,7 @@ func TestInspeqtorParse(t *testing.T) {
 	assert.Nil(t, err)
 	err = i.Parse()
 	assert.Nil(t, err)
-	assert.Equal(t, i.GlobalConfig.CycleTime, uint16(15))
+	assert.Equal(t, i.GlobalConfig.CycleTime, uint(15))
 	assert.Equal(t, i.GlobalConfig.Variables["foo"], "bar")
 	assert.Equal(t, 2, len(i.Services))
